+++ tests/unit/asyn/wallet/test_wallet.py
@@ -2,7 +2,6 @@ from unittest import TestCase
 
 from xrpl.asyncio.wallet.wallet_generation import (
     _DEV_FAUCET_URL,
-    _HOOKS_V3_TEST_FAUCET_URL,
     _TEST_FAUCET_URL,
     get_faucet_url,
     process_faucet_host_url,
@@ -38,14 +37,6 @@ class TestWallet(TestCase):
         json_client_url = "https://testnet.xrpl-labs.com"
         ws_client_url = "wss://testnet.xrpl-labs.com"
         expected_faucet = _TEST_FAUCET_URL
-
-        self.assertEqual(get_faucet_url(json_client_url), expected_faucet)
-        self.assertEqual(get_faucet_url(ws_client_url), expected_faucet)
-
-    def test_get_faucet_wallet_hooks_v3_test(self):
-        json_client_url = "https://hooks-testnet-v3.xrpl-labs.com"
-        ws_client_url = "wss://hooks-testnet-v3.xrpl-labs.com"
-        expected_faucet = _HOOKS_V3_TEST_FAUCET_URL
 
         self.assertEqual(get_faucet_url(json_client_url), expected_faucet)
         self.assertEqual(get_faucet_url(ws_client_url), expected_faucet)